Move new fields to CCDFields class
