+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/domain/model/CcdFields.java
@@ -22,5 +22,11 @@ public class CcdFields {
     public static final String GENERAL_ORDER_DETAILS = "GeneralOrderDetails";
     public static final String GENERAL_ORDER_DATE = "GeneralOrderDate";
     public static final String GENERAL_ORDER_RECITALS = "GeneralOrderRecitals";
+    public static final String GENERAL_EMAIL_DETAILS = "GeneralEmailDetails";
     public static final String CO_RESPONDENT_LINKED_TO_CASE = "CoRespLinkedToCase";
+    public static final String CO_RESPONDENT_SOLICITOR_EMAIL_ADDRESS = "CoRespondentSolicitorEmail";
+
+    // Other Party
+    public static final String GENERAL_EMAIL_OTHER_RECIPIENT_NAME = "GeneralEmailOtherRecipientName";
+    public static final String GENERAL_EMAIL_OTHER_RECIPIENT_EMAIL = "GeneralEmailOtherRecipientEmail";
 }
